+++ src/BaelorApi/Models/ViewModels/PatchLyricViewModel.cs
@@ -8,6 +8,6 @@ namespace BaelorApi.Models.ViewModels
 		public string Content { get; set; }
 
 		[JsonProperty("time_code")]
-		public int TimeCode { get; set; }
+		public float TimeCode { get; set; }
 	}
 }
